Specify make targets explicitly in nix package
The umu-vendored make target depends on git submodules.

Until now, it was disabled with a patch, however that broke in #330.

Instead of patching the makefile's `all` target, we can explicitly
specify the targets we want to build/install using make flags.

This is a temporary workaround until we can figure out how to get
submodules working correctly with the flake.

The flakeref query `submodules=1` _should_ fix this, but I couldn't get
it to work for some reason.